Change arc weight type from int to float for precise thresholds
All arc classes now accept float weights instead of int:
- Arc.__init__(weight: float = 1)
- Arc.set_weight(weight: float)
- InhibitorArc.__init__(weight: float = 1)
- TestArc.__init__(weight: float = 1)
- CurvedArc.__init__(weight: float = 1)
- CurvedInhibitorArc.__init__(weight: float = 1)

This allows precise threshold values like:
- Inhibitor arc weight = 2.5 mM (not just 2 or 3)
- Stoichiometric coefficients = 1.5 (e.g., fractional reactions)
- Test arc thresholds = 0.5 (activation at half concentration)

Example 08 already uses float weights (2.5, 2.0) in JSON,
now properly supported by type system.

+++ src/shypn/netobjs/arc.py
@@ -21,7 +21,7 @@ class Arc(PetriNetObject):
     DEFAULT_WIDTH = 3.0  # Legacy: 3.0px line width
     ARROW_SIZE = 15.0    # Legacy: 15px arrowhead length
     
-    def __init__(self, source, target, id: str, name: str, weight: int = 1):
+    def __init__(self, source, target, id: str, name: str, weight: float = 1):
         """Initialize an Arc.
         
         Args:
@@ -647,7 +647,7 @@ class Arc(PetriNetObject):
         """
         pass  # Arcs move when their source/target objects move
     
-    def set_weight(self, weight: int):
+    def set_weight(self, weight: float):
         """Set the arc weight.
         
         Args:
